save last known build dir and use if no other info is available
