+++ .gitignore
@@ -6,6 +6,7 @@ cmake-build*
 
 js/.idea
 js/node_modules
+js/cmake_binary_dir.txt
 
 js/api_tests/node_modules
 js/api_tests/tmp
@@ -13,7 +14,6 @@ js/api_tests/tmp
 js/client/dist
 js/client/node_modules
 
-js/webui/build_config
 js/webui/node_modules
 
 cpp/.idea
