+++ js/CMakeLists.txt
@@ -33,6 +33,8 @@ add_custom_command(
 
 add_custom_target(js_packages DEPENDS ${YARN_INTEGRITY})
 
+configure_file(cmake_binary_dir.txt.in ${CMAKE_CURRENT_SOURCE_DIR}/cmake_binary_dir.txt)
+
 if(ENABLE_TESTS)
     add_subdirectory(api_tests)
 endif()
