+++ js/api_tests/src/foobar2000/test_context.js
@@ -1,4 +1,5 @@
 import path from 'path';
+import { getBinaryDir } from '../../../config.mjs';
 import { TestContextFactory } from '../test_context.js';
 import PlayerController from './player_controller.js';
 
@@ -15,9 +16,7 @@ export class Foobar2000TestContextFactory extends TestContextFactory
         config.playerDir = path.join(config.appsDir, 'foobar2000', config.playerVersion);
 
         config.pluginBuildDir = path.join(
-            config.rootDir,
-            'build',
-            config.buildType,
+            getBinaryDir(config.buildType),
             'cpp',
             'server',
             'foobar2000',
