+++ js/cmake_binary_dir.txt.in
@@ -1,0 +1,1 @@
+@CMAKE_BINARY_DIR@
